more deserialization changes
