+++ src/LaunchDarkly.Client/LaunchDarkly.Client.csproj
@@ -74,7 +74,6 @@
     <Compile Include="Operator.cs" />
     <Compile Include="PollingProcessor.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ShaHex.cs" />
     <Compile Include="User.cs" />
     <Compile Include="FeatureFlag.cs" />
     <Compile Include="Util.cs" />
