+++ src/LaunchDarkly.Client/User.cs
@@ -13,24 +13,34 @@ namespace LaunchDarkly.Client
 
         [JsonProperty(PropertyName = "key", NullValueHandling = NullValueHandling.Ignore)]
         public string Key { get; set; }
+
         [JsonProperty(PropertyName = "secondary", NullValueHandling = NullValueHandling.Ignore)]
         public string SecondaryKey { get; set; }
+
         [JsonProperty(PropertyName = "ip", NullValueHandling = NullValueHandling.Ignore)]
         public string IpAddress { get; set; }
+
         [JsonProperty(PropertyName = "country", NullValueHandling = NullValueHandling.Ignore)]
         public string Country { get; set; }
+
         [JsonProperty(PropertyName = "firstName", NullValueHandling = NullValueHandling.Ignore)]
         public string FirstName { get; set; }
+
         [JsonProperty(PropertyName = "lastName", NullValueHandling = NullValueHandling.Ignore)]
         public string LastName { get; set; }
+
         [JsonProperty(PropertyName = "name", NullValueHandling = NullValueHandling.Ignore)]
         public string Name { get; set; }
+
         [JsonProperty(PropertyName = "avatar", NullValueHandling = NullValueHandling.Ignore)]
         public string Avatar { get; set; }
+
         [JsonProperty(PropertyName = "email", NullValueHandling = NullValueHandling.Ignore)]
         public string Email { get; set; }
-        [JsonProperty(PropertyName ="anonymous", NullValueHandling = NullValueHandling.Ignore)]
-        public  bool? Anonymous { get; set; }
+
+        [JsonProperty(PropertyName = "anonymous", NullValueHandling = NullValueHandling.Ignore)]
+        public bool? Anonymous { get; set; }
+
         [JsonProperty(PropertyName = "custom", NullValueHandling = NullValueHandling.Ignore)]
         public Dictionary<string, JToken> Custom { get; set; }
 
@@ -75,8 +85,8 @@ namespace LaunchDarkly.Client
                     return val.Value;
                     */
             }
-        }  
-        
+        }
+
         public User(string key)
         {
             Key = key;
@@ -86,26 +96,6 @@ namespace LaunchDarkly.Client
         public static User WithKey(string key)
         {
             return new User(key);
-        }
-
-        public float GetParam(string featureKey, string salt)
-        {
-            var idHash = Key;
-
-            if (!string.IsNullOrEmpty(SecondaryKey))
-                idHash += "." + SecondaryKey; 
-
-            var hash = ShaHex.Hash(string.Format("{0}.{1}.{2}", featureKey, salt, idHash)).Substring(0, 15);
-
-            var longValue = long.Parse(hash, NumberStyles.HexNumber);
-            const float longScale = 0xFFFFFFFFFFFFFFFL;
-
-            return longValue/longScale;
-        }
-
-        public JToken GetCustom(string attribute)
-        {
-            return Custom[attribute];
         }
 
     }
